pyproject.toml
[project]
name = "ai-changes-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "shopifyapi>=12.7.0",
]
